+++ hasher-matcher-actioner/hmalib/common/tests/config_test.py
@@ -278,46 +278,3 @@ class ConfigTest(unittest.TestCase):
         match_message = MatchMessage("key", "hash", [])
 
         assert perform_label_action(match_message, action_label) == 1
-
-    def test_react_action(self):
-        banked_signals = [
-            BankedSignal("2862392437204724", "bank 4", "te"),
-            BankedSignal("4194946153908639", "bank 4", "te"),
-        ]
-
-        configs: t.List[ActionPerformer] = [
-            ReactInReviewActionPerformer(
-                action_label=ActionLabel("ReactInReview"),
-            ),
-            ReactSawThisTooActionPerformer(
-                action_label=ActionLabel("ReactSawThisToo"),
-            ),
-        ]
-        for c in configs:
-            ActionPerformerConfig.update_performer(c)
-
-        action_label = ActionLabel("ReactInReview")
-
-        match_message = MatchMessage("key", "hash", banked_signals)
-        assert perform_label_action(match_message, action_label) == 1
-
-        action_label = ActionLabel("ReactSawThisToo")
-        assert perform_label_action(match_message, action_label) == 1
-
-    def test_react_non_te(self):
-        banked_signals = [
-            BankedSignal("whbrvohberov", "bank 4", "local bank"),
-            BankedSignal("4ytmnetn3ty", "bank 4", "local bank"),
-        ]
-
-        configs: t.List[ActionPerformer] = [
-            ReactInReviewActionPerformer(
-                action_label=ActionLabel("ReactInReview"),
-            ),
-        ]
-        for c in configs:
-            ActionPerformerConfig.update_performer(c)
-
-        action_label = ActionLabel("ReactInReview")
-        match_message = MatchMessage("key", "hash", banked_signals)
-        assert perform_label_action(match_message, action_label) == 1
